+++ engine/domain/README.md
@@ -1,0 +1,3 @@
+# Domain
+
+Pure, deterministic functions that operate on domain models. No game state, no player turn logic.
