+++ engine/reducers/index.ts
@@ -1,7 +1,7 @@
 export { addPlayerReducer, removePlayerReducer } from './playerReducers.ts';
 export { foundHotelReducer } from './foundHotelReducer.ts';
 export { startGameReducer } from './gameReducers.ts';
-export { resolveMergerReducer } from './resolveMergerReducer.ts';
-export { breakMergerTieReducer } from './breakMergerTieReducer.ts';
+export { completeMergerReducer } from './completeMergerReducer.ts';
+export { prepareMergerReducer } from './prepareMergerReducer.ts';
 export { buySharesReducer } from './buySharesReducer.ts';
 export { playTileReducer } from './playTileReducer.ts';
