+++ engine/usecases/index.ts
@@ -1,0 +1,4 @@
+export { advanceTurnUseCase } from './advanceTurnUseCase.ts';
+export { breakMergerTieUseCase } from './breakMergerTieUseCase.ts';
+export { playTileUseCase } from './playTileUseCase.ts';
+export { resolveMergerUseCase } from './resolveMergerUseCase.ts';
